Updated object import function and README instructions
The object import function in the Blender script has been modified to use a different operation.

The README file was also updated with more detailed instructions for setting up the environment, installing dependencies, running demo scripts, and executing Blender scripts.

A few corrections were made to the commands as well.